refactor: restructure assistant API to use a service layer for improved organization and maintainability
